pyproject.toml
[project]
name = "central-liveai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "google-genai>=1.70.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "websockets>=16.0",
]
